+++ packages/sensors/example/android/app/src/main/java/io/flutter/plugins/sensorsexample/MainActivity.java
@@ -1,13 +1,14 @@
 package io.flutter.plugins.sensorsexample;
 
-import android.os.Bundle;
-import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
+import io.flutter.embedding.android.FlutterActivity;
+import io.flutter.embedding.engine.FlutterEngine;
+import io.flutter.plugins.sensors.SensorsPlugin;
 
 public class MainActivity extends FlutterActivity {
+
   @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
+  public void configureFlutterEngine(FlutterEngine flutterEngine) {
+    super.configureFlutterEngine(flutterEngine);
+    flutterEngine.getPlugins().add(new SensorsPlugin());
   }
 }
